refactor(cli): split services command into platform-specific files
